backend/requirements.txt
# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Async SSE
sse-starlette>=1.8.0

# CORS
python-dotenv>=1.0.0

# Forward compatibility with existing modules
openai>=1.12.0
replicate>=0.23.0
requests>=2.31.0
edge-tts>=6.1.10
moviepy>=2.0.0
Pillow>=10.2.0
oss2>=2.18.0

